feat: settings
