+++ pages/index.vue
@@ -1,20 +1,8 @@
-<script setup lang="ts">
-import {storeToRefs} from "pinia";
-
-const stickerStore = useStickersStore();
-const { stickers } = storeToRefs(stickerStore)
-</script>
 <template>
-  <div class="dashboard-page">
-    <Sticker v-for="item in stickers" :item="item" />
+  <div class="normal-page">
+    <StickersView />
   </div>
 </template>
-
-<style lang="scss" scoped>
-.dashboard-page {
-  width: 100%;
-  display: flex;
-  gap: 20px;
-  flex-wrap: wrap;
-}
-</style>
+<script setup lang="ts">
+import StickersView from "~/components/StickersView.vue";
+</script>
