+++ scss/main.scss
@@ -9,3 +9,8 @@ html, body, #__nuxt {
   margin: 0;
   padding: 0;
 }
+
+.normal-page {
+  width: 100%;
+  height: 100%;
+}
